finished off getClosest

+++ core/modules/vehicle/client.lua
@@ -1,4 +1,5 @@
 QBCorev2.Modules.Vehicle = {}
+--- Spawns a vehicle at the specified coordinates and returns the vehicle and networkid, client-side only
 ---@param model string|number -- model name or hash
 ---@param cb function -- callback function
 ---@param coords vector4|false -- vector4 or false
@@ -32,6 +33,7 @@ function QBCorev2.Modules.Vehicle.Spawn(model, cb, coords, isnetworked, teleport
 end
 
 function QBCorev2.Modules.Vehicle.Delete(vehicle)
+    if not DoesEntityExist(vehicle) then return error('this entity does not exist: '..vehicle, 2)end
     SetEntityAsMissionEntity(vehicle, true, true)
     DeleteVehicle(vehicle)
 end
@@ -78,4 +80,21 @@ function QBCorev2.Modules.Vehicle.GetProperties(vehicle)
         props.mods[i] = GetVehicleMod(vehicle, i)
     end
     return props
+end
+
+-- backwards compatibility
+function QBCore.Functions.SpawnVehicle(model, cb, coords, isnetworked, teleportInto)
+    QBCorev2.Modules.Vehicle.Spawn(model, cb, coords, isnetworked, teleportInto)
+end
+
+function QBCore.Functions.DeleteVehicle(vehicle)
+    QBCorev2.Modules.Vehicle.Delete(vehicle)
+end
+
+function QBCore.Functions.GetVehicleProperties(vehicle)
+    QBCorev2.Modules.Vehicle.GetProperties(vehicle)
+end
+
+function QBCore.Functions.GetVehiclePlate(vehicle)
+    QBCorev2.Modules.Vehicle.GetPlate(vehicle)
 end
